// web/src/api/generated/models/catalogSetItemDetailsRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * POS Cafe API
 * Backend API for Cafe Point of Sale System built with Vertical Slice Architecture and Idiomatic Go.
 * OpenAPI spec version: 1.0
 */
import type { CatalogSetItemDetailsRequestBadge } from './catalogSetItemDetailsRequestBadge';

export interface CatalogSetItemDetailsRequest {
  badge?: CatalogSetItemDetailsRequestBadge;
  code?: string;
  description?: string;
  request_id?: string;
}